Adds support for python modules
Sources of python modules go in a folder in src/python.
During the build they are installed into a folder with the same name in
lib/python and installed in the same folder in $AIS_ROOT/lib/python.
Note that python scripts (i.e. files in src/python/*.py) are not part of
modules but executables and are, in fact, installed in ../bin.