Moves file methods in FileSupport and build the classpath
Moves the methods to deal with files and folders, as well as the
definition of the IAS folders from CommonDefs into FileSupport.
CommonDefs provides a method to build the classpath for running java and
scala executables (probably not th ebast place or name for a class but
it will be ok for now).