Acknowledgement state of an alarm is a property.
AlarmValue.acknowledgement says if the operator acknowledged the alarm.
This logic  was previously implemented by 2 alarm states ActiveAndNew
and ActiveAndAcknowledged but there was only a Cleared stated i.e. the
ClearedAndNotAcknowledged case was not handled. This let me think that
an alarm is acknowledged or new independently of its actual state.