Sets default java/scala properties
Set defaults java/scala properties i.e. properties that must be defined
to all executables. An example can confiduration properties.
The future can easily be extended for the user to set other properties
eventually overriding the defaults. 

iasRun prints a message with the java/scala command is going to run.