MonitorPoint factory methods moved into TypedMonitorPoint
Factory methods have been moved out of the MonitorPoint into the new
TypedMonitorPoint. Type checking on the implemented types is done at
build time in case we wish to keep types under control and provide new
specialized types.

MonitorPoint[A] is now merely a value holder for values of type A
extending MonitorPointBase so that we do not have to bring the type
around while dealing with different types of monitor points.

Terst updated accordingly