Initial version of the Integrated Alarm System Component.
The ASC is the component that digest the inputs to produce the output.
At the present it handles the different input types, shelving and
acknowledging of alarms but does not yet run any script to generate the
output ot of the values of the inputs.
I am still evaluating different solutions for that.