Replaces parametrized MonitorPoints with abstract types and vals
Refactors parametrized MonitorPoints using abstract types and variables.
In this way we got rid of several classes in BasicTypes and the codes is
shorter because we do not have to refer to a generic un-parametrized
base class. MonitorPointFactory has been removed as well and the factory
moved into MonitorPoint itself.

We will see later if this refactoring has been a good idea and if it is
the case to adopt the same strategy for the ASCE in favor of the DASU.

The drawback of this strategy is that type checking while updating a
value is done at run-time. For this reason the type checking of the
value of the monitor points has been reinforced all around the code.
While we can trust what happens in the core, we have to be very careful
on the user implemented transfer functions.

Previous version tagged 0.0.1.