HeteroInOut refactored
HIO has been refactored to get rid of type paremeters and to make its
useage more friendly.
There is only on class now, named InOut[T]. The actualValue is always
there but it is a Option in case it has not yet been set or its value
became invalid.
Best gain IMO is the type checking and compile time.

Tests pass but other module stil needs to be fixed.