The ASCE is abstract and the transfer is mixed in
The ASC is abstract with the transfer(...) method undefined.
The transfer method depends on the prograaming language of the user
provided implementation and is mixed in when building the ASCE itslef.
For example if the programaming language is scala the ASCE is built with
a command like:
  new ComputingElement[T](...) with ScalaTransfer[T]
  
This ensures that only one implementation is given and shorten the
execution time in respect of stackable modifications used before.